Backend/upload-report/requirements.txt
fastapi
uvicorn
google-cloud-storage
google-cloud-firestore
python-multipart